requirements.txt
requests>=2.21.0
veracode-api-py>=0.9.10
logging-formatter-anticrlf>=1.2